+++ soh/soh/Enhancements/game-interactor/GameInteractor.h
@@ -266,6 +266,8 @@ typedef enum {
     GI_VB_GIVE_ITEM_FROM_MEDIGORON,
     // Opt: *EnMs
     GI_VB_GIVE_ITEM_FROM_MAGIC_BEAN_SALESMAN,
+    // Opt: *EnFr
+    GI_VB_GIVE_ITEM_FROM_FROGS,
 
     GI_VB_GIVE_ITEM_FAIRY_OCARINA,
     GI_VB_GIVE_ITEM_WEIRD_EGG,
