+++ soh/src/overlays/actors/ovl_En_Fr/z_en_fr.c
@@ -45,7 +45,6 @@ void EnFr_OcarinaMistake(EnFr* this, PlayState* play);
 void EnFr_SetupReward(EnFr* this, PlayState* play, u8 unkCondition);
 void EnFr_PrintTextBox(EnFr* this, PlayState* play);
 void EnFr_TalkBeforeReward(EnFr* this, PlayState* play);
-RandomizerCheck EnFr_RandomizerCheckFromSongIndex(u16 songIndex);
 void EnFr_SetReward(EnFr* this, PlayState* play);
 
 // Deactivate
@@ -618,7 +617,6 @@ void EnFr_Idle(EnFr* this, PlayState* play) {
         player->actor.world.pos.z = this->actor.world.pos.z; // z = -1220.0f
         player->currentYaw = player->actor.world.rot.y = player->actor.shape.rot.y = this->actor.world.rot.y;
         this->reward = GI_NONE;
-        this->getItemEntry = (GetItemEntry)GET_ITEM_NONE;
         this->actionFunc = EnFr_Activate;
     } else if (EnFr_IsAboveAndWithin30DistXZ(player, this)) {
         player->unk_6A8 = &this->actor;
@@ -854,7 +852,6 @@ s32 EnFr_IsFrogSongComplete(EnFr* this, PlayState* play) {
 void EnFr_OcarinaMistake(EnFr* this, PlayState* play) {
     Message_CloseTextbox(play);
     this->reward = GI_NONE;
-    this->getItemEntry = (GetItemEntry)GET_ITEM_NONE;
     func_80078884(NA_SE_SY_OCARINA_ERROR);
     Audio_OcaSetInstrument(0);
     sEnFrPointers.flags = 12;
@@ -939,23 +936,6 @@ void EnFr_TalkBeforeReward(EnFr* this, PlayState* play) {
     }
 }
 
-RandomizerCheck EnFr_RandomizerCheckFromSongIndex(u16 songIndex) {
-    switch (songIndex) {
-        case FROG_ZL:
-            return RC_ZR_FROGS_ZELDAS_LULLABY;
-        case FROG_EPONA:
-            return RC_ZR_FROGS_EPONAS_SONG;
-        case FROG_SARIA:
-            return RC_ZR_FROGS_SARIAS_SONG;
-        case FROG_SUNS:
-            return RC_ZR_FROGS_SUNS_SONG;
-        case FROG_SOT:
-            return RC_ZR_FROGS_SONG_OF_TIME;
-        default:
-            return RC_UNKNOWN_CHECK;
-    }
-}
-
 void EnFr_SetReward(EnFr* this, PlayState* play) {
     u16 songIndex;
 
@@ -963,16 +943,12 @@ void EnFr_SetReward(EnFr* this, PlayState* play) {
     songIndex = this->songIndex;
     this->actionFunc = EnFr_Deactivate;
     this->reward = GI_NONE;
-    this->getItemEntry = (GetItemEntry)GET_ITEM_NONE;
     if ((songIndex >= FROG_ZL) && (songIndex <= FROG_SOT)) {
         if (!(gSaveContext.eventChkInf[13] & sSongIndex[songIndex])) {
             gSaveContext.eventChkInf[13] |= sSongIndex[songIndex];
             GameInteractor_ExecuteOnFlagSet(FLAG_EVENT_CHECK_INF, (EVENTCHKINF_SONGS_FOR_FROGS_INDEX << 4) + sSongIndexShift[songIndex]);
-            if (!IS_RANDO) {
+            if (GameInteractor_Should(GI_VB_GIVE_ITEM_FROM_FROGS, true, this)) {
                 this->reward = GI_RUPEE_PURPLE;
-            } else {
-                this->getItemEntry = Randomizer_GetItemFromKnownCheck(EnFr_RandomizerCheckFromSongIndex(songIndex), GI_RUPEE_PURPLE);
-                this->reward = this->getItemEntry.getItemId;
             }
         } else {
             this->reward = GI_RUPEE_BLUE;
@@ -981,11 +957,8 @@ void EnFr_SetReward(EnFr* this, PlayState* play) {
         if (!(gSaveContext.eventChkInf[13] & sSongIndex[songIndex])) {
             gSaveContext.eventChkInf[13] |= sSongIndex[songIndex];
             GameInteractor_ExecuteOnFlagSet(FLAG_EVENT_CHECK_INF, (EVENTCHKINF_SONGS_FOR_FROGS_INDEX << 4) + sSongIndexShift[songIndex]);
-            if (!IS_RANDO) {
+            if (GameInteractor_Should(GI_VB_GIVE_ITEM_FROM_FROGS, true, this)) {
                 this->reward = GI_HEART_PIECE;
-            } else {
-                this->getItemEntry = Randomizer_GetItemFromKnownCheck(RC_ZR_FROGS_IN_THE_RAIN, GI_HEART_PIECE);
-                this->reward = this->getItemEntry.getItemId;
             }
         } else {
             this->reward = GI_RUPEE_BLUE;
@@ -994,11 +967,8 @@ void EnFr_SetReward(EnFr* this, PlayState* play) {
         if (!(gSaveContext.eventChkInf[13] & sSongIndex[songIndex])) {
             gSaveContext.eventChkInf[13] |= sSongIndex[songIndex];
             GameInteractor_ExecuteOnFlagSet(FLAG_EVENT_CHECK_INF, (EVENTCHKINF_SONGS_FOR_FROGS_INDEX << 4) + sSongIndexShift[songIndex]);
-            if (!IS_RANDO) {
+            if (GameInteractor_Should(GI_VB_GIVE_ITEM_FROM_FROGS, true, this)) {
                 this->reward = GI_HEART_PIECE;
-            } else {
-                this->getItemEntry = Randomizer_GetItemFromKnownCheck(RC_ZR_FROGS_OCARINA_GAME, GI_HEART_PIECE);
-                this->reward = this->getItemEntry.getItemId;
             }
         } else {
             this->reward = GI_RUPEE_PURPLE;
@@ -1049,29 +1019,23 @@ void EnFr_Deactivate(EnFr* this, PlayState* play) {
         this->actionFunc = EnFr_Idle;
     } else {
         this->actionFunc = EnFr_GiveReward;
-        if (!IS_RANDO || this->getItemEntry.getItemId == GI_NONE) {
+        if (GameInteractor_Should(GI_VB_GIVE_ITEM_FROM_FROGS, true, this)) {
             func_8002F434(&this->actor, play, this->reward, 30.0f, 100.0f);
-        } else {
-            GiveItemEntryFromActor(&this->actor, play, this->getItemEntry, 30.0f, 100.0f);
         }
     }
 }
 
 void EnFr_GiveReward(EnFr* this, PlayState* play) {
-    if (Actor_HasParent(&this->actor, play)) {
+    if (Actor_HasParent(&this->actor, play) || !GameInteractor_Should(GI_VB_GIVE_ITEM_FROM_FROGS, true, this)) {
         this->actor.parent = NULL;
         this->actionFunc = EnFr_SetIdle;
     } else {
-        if (!IS_RANDO || this->getItemEntry.getItemId == GI_NONE) {
-            func_8002F434(&this->actor, play, this->reward, 30.0f, 100.0f);
-        } else {
-            GiveItemEntryFromActor(&this->actor, play, this->getItemEntry, 30.0f, 100.0f);
-        }
+        func_8002F434(&this->actor, play, this->reward, 30.0f, 100.0f);
     }
 }
 
 void EnFr_SetIdle(EnFr* this, PlayState* play) {
-    if ((Message_GetState(&play->msgCtx) == TEXT_STATE_DONE) && Message_ShouldAdvance(play) || (IS_RANDO && this->reward == RG_ICE_TRAP)) {
+    if ((Message_GetState(&play->msgCtx) == TEXT_STATE_DONE) && Message_ShouldAdvance(play)) {
         this->actionFunc = EnFr_Idle;
     }
 }
